test(runners): isolate token-command security coverage
